requirements.txt
aiogram==2.25.1
aiosqlite
gspread
google-auth
